requirements.txt
librosa
whisper
pysrt
six
ffmpeg-python
torch